Reduce redundant code
