+++ src/components/screen/ProfileUpdate.tsx
@@ -1,17 +1,19 @@
-import * as ImagePicker from 'expo-image-picker';
-import * as Permissions from 'expo-permissions';
 import React, { useEffect, useState } from 'react';
 import { TouchableOpacity, View } from 'react-native';
-
+import {
+  launchCameraAsync,
+  launchImageLibraryAsync,
+} from '../../utils/ImagePicker';
 import Button from '../shared/Button';
-// import { CommonActions } from '@react-navigation/core';
 import { DefaultNavigationProps } from '../../types';
 import { Ionicons } from '@expo/vector-icons';
+
 import TextInput from '../shared/TextInput';
 import { getString } from '../../../STRINGS';
 import styled from 'styled-components/native';
 import { useActionSheet } from '@expo/react-native-action-sheet';
 import { useThemeContext } from '../../providers/ThemeProvider';
+// import { CommonActions } from '@react-navigation/core';
 
 const StyledContainer = styled.View`
   flex: 1;
@@ -91,25 +93,13 @@ function Screen(props: Props): React.ReactElement {
     }
   };
 
-  const getPermissions = async (
-    type: string,
-  ): Promise<Permissions.PermissionStatus> => {
-    if (type === 'photo') {
-      const { status } = await Permissions.askAsync(Permissions.CAMERA_ROLL);
-      return status;
-    }
-    const { status } = await Permissions.askAsync(Permissions.CAMERA);
-    return status;
-  };
-
   const onPressImg = (): void => {
-    const options = ['촬영하기', '앨범에서 선택하기', '취소'];
+    const options = [
+      getString('TAKE_A_PICTURE'),
+      getString('SELSCT_FROM_ALBUM'),
+      getString('CANCEL'),
+    ];
     const cancelButtonIndex = 2;
-    const photoOptions = {
-      mediaTypes: ImagePicker.MediaTypeOptions.Images,
-      allowsEditing: true,
-      exif: true,
-    };
 
     showActionSheetWithOptions(
       {
@@ -118,20 +108,16 @@ function Screen(props: Props): React.ReactElement {
       },
       async (buttonIndex: number) => {
         if (buttonIndex === 0) {
-          const permissionStatus = await getPermissions('camera');
-          if (permissionStatus === Permissions.PermissionStatus.GRANTED) {
-            const result = await ImagePicker.launchCameraAsync(photoOptions);
+          const result = await launchCameraAsync();
+          if (result) {
             console.log('result', result);
           }
           return;
         }
 
         if (buttonIndex === 1) {
-          const permissionStatus = await getPermissions('photo');
-          if (permissionStatus === Permissions.PermissionStatus.GRANTED) {
-            const result = await ImagePicker.launchImageLibraryAsync(
-              photoOptions,
-            );
+          const result = await launchImageLibraryAsync();
+          if (result) {
             console.log('result', result);
           }
         }
